requirements.txt
fastapi
uvicorn
google-cloud-storage
pydantic-settings
Pillow
reportlab
